+++ ndexncipidloader/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Chris Churas"""
 __email__ = 'contact@ndexbio.org'
-__version__ = '1.5.1'
+__version__ = '1.6.0'